Combine PIRIngestor and PIRLinker
